Move db-discovery statement to db-discovery module
